refactor: move activity.js
